requirements.txt
flask
torch
torchvision
pillow
scikit-image
lpips
numpy
